Added "all" devices, fixed line breaks.
I added the option to send a notice to all devices by overriding the device ID with the API ID it finds.  Only works to send to your own devices, however, not someone elses.  Though in theory if you specify the email address it might work to send to someone else's devices.

Fixed line breaks in the help text so it is formatted better for a terminal with the standard 80 character width.